+++ cross/apple/Demo/App/Configuration/Demo+Profile.swift
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import Foundation
-import PartoutCore
+import PartoutRuntime
 
 extension Profile {
     static let demo: Profile = {
