+++ cross/apple/Demo/App/DemoApp.swift
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-import PartoutCore
+import PartoutRuntime
 import SwiftUI
 
 @main
