+++ cross/apple/Demo/Shared/Shared+App.swift
@@ -2,20 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-import Foundation
-import PartoutCore
-import PartoutNative
-
-extension Demo {
-    static func parseModule<M>(_ type: M.Type, url: URL) throws -> M? where M: Decodable {
-        let text = try String(contentsOf: url, encoding: .utf8)
-        guard let cJSON = partout_import_module(text) else { return nil }
-        let json = String(cString: cJSON)
-        guard let jsonData = json.data(using: .utf8) else { return nil }
-        let tagged = try JSONDecoder.shared().decode(TaggedModule.self, from: jsonData)
-        return tagged.containedModule as? M
-    }
-}
+import PartoutRuntime
 
 extension Demo.Log {
     static let appURL = Demo.cachesURL.appending(component: "app.log")
