--- cross/apple/Sources/PartoutCore/PartoutConstants.swift
@@ -1,20 +1,0 @@
-// SPDX-FileCopyrightText: 2026 Davide De Rosa
-//
-// SPDX-License-Identifier: GPL-3.0
-
-public enum PartoutConstants {
-    /// The unique identifier of the library.
-    public static let identifier = "io.partout"
-
-    /// The library version.
-    public static let version = "0.155.2"
-
-    /// The computed version identifier.
-    public static let versionIdentifier: String = "\(identifier) \(version)"
-
-    /// The C flavor of ``versionIdentifier``.
-    public static var cVersionIdentifier: UnsafePointer<CChar> {
-        // This is safe because the subject is statically allocated.
-        versionIdentifier.withCString(\.self)
-    }
-}
