+++ cross/apple/Sources/PartoutCore/PartoutCore.swift
@@ -2,10 +2,25 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-#if !USE_CMAKE
-@_exported import Foundation
 @_exported import _PartoutPortable_C
-#endif
+@_exported import Foundation
+
+public enum PartoutConstants {
+    /// The unique identifier of the library.
+    public static let identifier = "io.partout"
+
+    /// The library version.
+    public static let version = "0.155.2"
+
+    /// The computed version identifier.
+    public static let versionIdentifier: String = "\(identifier) \(version)"
+
+    /// The C flavor of ``versionIdentifier``.
+    public static var cVersionIdentifier: UnsafePointer<CChar> {
+        // This is safe because the subject is statically allocated.
+        versionIdentifier.withCString(\.self)
+    }
+}
 
 extension LoggerCategory: CaseIterable {
     public static let abi = Self(rawValue: "abi")
