+++ cross/apple/Sources/PartoutRuntime/PartoutRuntime.swift
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2026 Davide De Rosa
+//
+// SPDX-License-Identifier: GPL-3.0
+
+@_exported import PartoutCore
