+++ cross/apple/Sources/PartoutRuntime/PartoutTunnelController.swift
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import NetworkExtension
-import PartoutCore
 import PartoutNative
 
 /// A controller based on `NEPacketTunnelProvider`.
